% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bac_tax.R
\docType{data}
\name{bac_tax}
\alias{bac_tax}
\title{Bacterial taxonomy information}
\format{
A data frame with rows for each bacterial OTU and columns:
\describe{
  \item{id}{Bacterial OTU ID}
  \item{Kingdom}{Kingdom}
  \item{Phylum}{Phylum classification}
  \item{Class}{Class classification}
  \item{Order}{Order classification}
  \item{Family}{Family classification}
  \item{Genus}{Genus classification}
  \item{Species}{Species classification}
}
}
\source{
\url{https://example.com}
}
\usage{
data(bac_tax)
}
\description{
A dataset containing taxonomic classification for bacteria.
}
\keyword{datasets}
