% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bacteria_data.R
\docType{data}
\name{bacteria_data}
\alias{bacteria_data}
\title{Bacteria abundance data}
\format{
A numeric matrix with rows representing bacterial OTUs and columns representing samples.
}
\source{
\url{https://example.com}
}
\usage{
bacteria_data
}
\description{
A matrix of bacterial abundances with bacterial OTUs as rows and samples as columns.
}
\keyword{datasets}
