% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votu_functions.R
\name{get_direction_info}
\alias{get_direction_info}
\title{Query records for a specific enrichment direction}
\usage{
get_direction_info(
  direction = c("Enriched_in_PD", "Enriched_in_HC"),
  data = NULL
)
}
\arguments{
\item{direction}{"Enriched_in_PD" or "Enriched_in_HC"}

\item{data}{Optional data frame}
}
\value{
Data frame
}
\description{
Query records for a specific enrichment direction
}
\examples{
\dontrun{
get_direction_info("Enriched_in_PD")
}
}
