% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votu_functions.R
\name{get_family_info}
\alias{get_family_info}
\title{Query all records for a specific phage family}
\usage{
get_family_info(family_name, data = NULL)
}
\arguments{
\item{family_name}{Family name}

\item{data}{Optional data frame}
}
\value{
Data frame
}
\description{
Query all records for a specific phage family
}
\examples{
\dontrun{
get_family_info("Siphoviridae")
}
}
