% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votu_functions.R
\name{get_ko_info}
\alias{get_ko_info}
\title{Query records containing a specific KO number}
\usage{
get_ko_info(ko_id, data = NULL, exact = FALSE)
}
\arguments{
\item{ko_id}{KO number}

\item{data}{Optional data frame}

\item{exact}{Logical; if TRUE, exact match; if FALSE, fuzzy match (default)}
}
\value{
Data frame
}
\description{
Query records containing a specific KO number
}
\examples{
\dontrun{
get_ko_info("K00001")
}
}
