% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votu_functions.R
\name{get_pathway_info}
\alias{get_pathway_info}
\title{Query records containing a specific pathway description}
\usage{
get_pathway_info(pathway_name, data = NULL, exact = FALSE)
}
\arguments{
\item{pathway_name}{Pathway description}

\item{data}{Optional data frame}

\item{exact}{Logical; if TRUE, exact match; if FALSE, fuzzy match (default)}
}
\value{
Data frame
}
\description{
Query records containing a specific pathway description
}
\examples{
\dontrun{
get_pathway_info("Glycolysis")
}
}
