% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votu_functions.R
\name{get_significant_votus}
\alias{get_significant_votus}
\title{Get list of significantly differential vOTUs}
\usage{
get_significant_votus(
  p_threshold = 0.05,
  data = NULL,
  adj = TRUE,
  return_records = FALSE
)
}
\arguments{
\item{p_threshold}{p-value threshold}

\item{data}{Optional data frame}

\item{adj}{Logical; use adjusted p-value (default TRUE)}

\item{return_records}{Logical; if TRUE, return full data frame; if FALSE, return vector of vOTU IDs}
}
\value{
Character vector or data frame
}
\description{
Get list of significantly differential vOTUs
}
\examples{
\dontrun{
get_significant_votus(p_threshold = 0.01)
}
}
