% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votu_functions.R
\name{get_votu_info}
\alias{get_votu_info}
\title{Query all records for a specific vOTU}
\usage{
get_votu_info(votu_id, data = NULL)
}
\arguments{
\item{votu_id}{vOTU ID}

\item{data}{Optional data frame}
}
\value{
Data frame
}
\description{
Query all records for a specific vOTU
}
\examples{
\dontrun{
get_votu_info("vOTU_123")
}
}
