% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-interaction_edges.R
\docType{data}
\name{interaction_edges_HC}
\alias{interaction_edges_HC}
\title{Bacteria-phage interaction edges for HC group}
\format{
A data frame with columns:
\describe{
  \item{Source}{Bacterial OTU ID}
  \item{Target}{vOTU ID}
  \item{Correlation}{Correlation value}
  \item{interaction}{Interaction type (if applicable)}
}
}
\source{
\url{https://example.com}
}
\usage{
data(interaction_edges_HC)
}
\description{
A data frame of edges (interactions) between bacteria and phages in the HC group.
}
\keyword{datasets}
