% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDmicroProfiler-package.R
\docType{package}
\name{PDmicroProfiler-package}
\alias{PDmicroProfiler}
\alias{PDmicroProfiler-package}
\title{PDmicroProfiler: Microbiome Profiler for Parkinson's Disease Research}
\description{
A package for analyzing and visualizing microbiome data in Parkinson's disease research, including abundance analysis, KO occurrence visualization, and bacteriophage-host interaction networks.
}
\author{
\strong{Maintainer}: Your Name \email{your.email@example.com}

}
\keyword{internal}
