% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votu_functions.R
\name{summarize_votu_function}
\alias{summarize_votu_function}
\title{Summarize functional information for a single vOTU}
\usage{
summarize_votu_function(votu_id, data = NULL)
}
\arguments{
\item{votu_id}{vOTU ID}

\item{data}{Optional data frame}
}
\value{
List with summary statistics
}
\description{
Summarize functional information for a single vOTU
}
\examples{
\dontrun{
summarize_votu_function("vOTU_123")
}
}
