% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-virus_data.R
\docType{data}
\name{virus_data}
\alias{virus_data}
\title{Virus abundance data}
\format{
A numeric matrix with rows representing vOTUs and columns representing samples.
}
\source{
\url{https://example.com}
}
\usage{
virus_data
}
\description{
A matrix of virus abundances with vOTUs as rows and samples as columns.
}
\keyword{datasets}
